ENH: create a 'mtasks' module for performing the 'tasks' in the primary import script.  Start moving code there.
