+++ scripts/import_sn/mtasks/__init__.py
@@ -1,0 +1,2 @@
+"""Individual (semi) operations for data importing.
+"""
